changed class name + all in 1 file
