+++ genlm/control/potential/__init__.py
@@ -5,7 +5,7 @@ from .operators import PotentialOps
 from .product import Product
 from .coerce import Coerced
 
-from .built_in import PromptedLLM, WCFG, BoolCFG, WFSA, BoolFSA, JsonSchema, Canonical
+from .built_in import PromptedLLM, WCFG, BoolCFG, WFSA, BoolFSA, JsonSchema, CanonicalTokenization
 
 __all__ = [
     "Potential",
@@ -17,7 +17,7 @@ __all__ = [
     "BoolCFG",
     "WFSA",
     "BoolFSA",
-    "Canonical",
+    "CanonicalTokenization",
     "AutoBatchedPotential",
     "MultiProcPotential",
     "Coerced",
