+++ genlm/control/potential/built_in/__init__.py
@@ -2,7 +2,7 @@ from .llm import PromptedLLM
 from .wcfg import WCFG, BoolCFG
 from .wfsa import WFSA, BoolFSA
 from .json import JsonSchema
-from .canonical import Canonical
+from .canonical import CanonicalTokenization
 
 __all__ = [
     "PromptedLLM",
@@ -11,5 +11,5 @@ __all__ = [
     "BoolCFG",
     "WFSA",
     "BoolFSA",
-    "Canonical",
+    "CanonicalTokenization",
 ]
